feat: organize logs
