+++ components/dashboard/logging/DashboardLoggingConfig.tsx
@@ -12,14 +12,8 @@ export default function DashboardLoggingConfig({ id }: { id: string }) {
         <div className={"animate-fadeIn flex max-md:items-center flex-col py-5 md:px-5 gap-5"}>
         <h1 className={"header text-6xl max-md:text-5xl"}>logging</h1>
         <span className={"flex max-md:flex-col max-w-[100vw] w-full gap-5 max-md:px-2"}>
-        <Suspense fallback={null}>
             <LogSettings server={logging?.data}/>
-        </Suspense>  
-        <Suspense fallback={null}>
-                <Logs logs={logging?.data?.logs}/>
-        </Suspense>   
-            
-            
+            <Logs logs={logging?.data?.logs}/> 
         </span>
         </div>
