+++ components/dashboard/logging/FilteredLogs.tsx
@@ -1,38 +1,13 @@
 "use client";
 
-
-
-import DashboardActionContext from "@/context/DashboardActionContext";
-
 import Link from "next/link";
-import { Suspense, useContext, useState } from 'react';
-import { BsCheckLg, BsPlus, BsX } from "react-icons/bs";
+import { Suspense, useState } from 'react';
+import { BsCheckLg, BsPlus } from "react-icons/bs";
 import { useQuery, useQueryClient } from "react-query";
 import { LogCombobox } from "./combobox/LogCombobox";
 import { Button } from "@/components/ui/button";
 import { useToast } from "@/components/ui/use-toast";
-
-function LogFilter({ filtered, serverID }: { filtered: string, index: number, serverID: string }) {
-
-    const { toast } = useToast();
-    const queryClient = useQueryClient();
-    function deleteReaction() {
-        if (!serverID) return;
-        const body = new URLSearchParams();
-        body.append("log", filtered);
-        fetch(`/api/v1/servers/${serverID}/log/filter`, { method: "POST", body }).then(async (data) => {
-            const json = await data.json().catch(() => undefined);
-            if (!json || json['error']) {
-                toast({ title: "Failed to remove filtered log", description: json['error'] ? json['error'] : 'An error occured.', status: "error" });
-                return;
-            }
-            queryClient.invalidateQueries(["data_logging", serverID])
-            toast({ title: "Filtered Log Removed", description: `Successfully removed "${filtered.split('_').map((i) => i[0] + i.slice(1).toLowerCase()).join(' ')}" as a filtered log.`, status: "success" })
-        }).catch(() => {     
-        });
-    }
-    return (<span className={"flex flex-row w-full justify-center gap-2 text-base items-center"}>{filtered.split('_').map((i) => i[0] + i.slice(1).toLowerCase()).join(' ')} <span className={"border text-white rounded-2xl w-fit h-fit p-1 hover-gradient transition-all hover:text-black hover:border-black text-sm cursor-pointer"} onClick={() => deleteReaction()}><BsX/></span></span>);
-}
+import { LogFilter } from "./LogFilter";
 
 export default function FilteredLogs({ server }: { server: { 
     serverID: string,
