+++ components/dashboard/logging/combobox/LogCombobox.tsx
@@ -1,11 +1,10 @@
 import { Button } from "@/components/ui/button";
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem } from "@/components/ui/command";
 import { Drawer, DrawerContent, DrawerTrigger } from "@/components/ui/drawer";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { cn } from "@/lib/utils";
-import { Check, ChevronsUpDown } from "lucide-react";
+import { ChevronsUpDown } from "lucide-react";
 import { useState } from "react";
 import { useMediaQuery } from "react-responsive";
+import { LogActionList } from "./LogActionList";
 
 interface LogProps {
     actions?: string[];
@@ -54,32 +53,4 @@ export function LogCombobox({ actions, value, onChange }: LogProps) {
           <LogActionList actions={actions} value={value} setOpen={setOpen} onChange={onChange} />
         </DrawerContent>
       </Drawer>)
-}
-function LogActionList({ actions, value, setOpen, onChange }: { actions?: string[], value: string, setOpen: (open: boolean) => void, onChange: (value: string) => void }) {
-    return (<Command value={value}>
-        <CommandInput placeholder="Search log actions..." />
-        <CommandEmpty>No log action found.</CommandEmpty>
-        <CommandGroup className={"max-h-[300px]"}>
-          {actions?.map((action) => (
-            <CommandItem
-              key={action}
-              value={action}
-              className={"flex items-center gap-2"}
-              onSelect={(currentValue) => {
-                onChange(currentValue === value ? "" : currentValue.toUpperCase())
-                setOpen(false)
-              }}
-            >
-              {action.split('_').map((i) => i.charAt(0).toUpperCase() + i.slice(1).toLowerCase()).join(' ')}
-              <Check
-                className={cn(
-                  "mr-2 h-4 w-4",
-                  value === action ? "opacity-100" : "opacity-0"
-                )}
-              />
-
-            </CommandItem>
-          ))}
-        </CommandGroup>
-      </Command>)
 }
